pyproject.toml
[tool.poetry]
name = "ml-kubernetes-eks"
version = "0.1.0"
description = ""
authors = ["HichamBenbriqa <hichambenbriqa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
scikit-learn = "1.2.2"
joblib = "^1.3.2"
neptune = "1.9.1"
python-dotenv = "1.0.0"
Flask = "3.0.2"
gunicorn = "21.2.0"
urllib3 = "1.26.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
ruff = "^0.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
